Add data processing environment
The data processing for this project will be modest, but we can
anticipate period data updates. Include an environment to isolate
dependencies and process the Excel files into a JSON format expected by
the application. The end result files generated by this commit are not
representative of the actual data expected, but demonstrate the
end-to-end functionality of the system.