Add data and config payloads to app
Add the generated config directly to the source code so it is in the
application bundle. As a result, we'll be able to render the homepage
immediately, without at async fetch.

Make the survey data accessible to fetch requests. This is not necessary
for immediate rendering, so can be deferred after initial page load.