Parse out questions/responses for visualization
An initial pass at a format to provide question/response pairs based on
visualization type for the selected question. It is anticipated that
additional shaping will be required for certain types, but this lays the
pathway with a focus on accounting for the "stack" type which aggregates
a number of variables from a single question.