Add Breadcrumbs
Adds breadcrumbs to the Question and Visualization components.
There's some extra logic to allow us to easily hook them in to
the Saved Charts page when we've built it.

While it might have been possible to create dynamically generated
breadcrumbs, making it a bit more declaratory simplified the code
significantly. Similarly, it would probably be possible to locate this
at the top level instead of in the specific components where it appears,
but because we are only going to be using it on three screens, the
extra logic seemed like an over-complication.

To keep the breadcrumbs and the route paths coordinated, I moved the
route paths to a ROUTES object in the constants file.