Add StackedBarChart
Adds the stacked bar chart, which is used for displaying agree/disagree
style survey questions. The legend and the country are included in the
canvas element, but it doesn't seem to be possible to include a title
for a canvas chart with Nivo. Hypothetically we could include the
question as an axis label (see the way that the country is included),
but there is no text wrapping applied to the axis labels, and the
questions are significantly longer than the space available.