Clear question state before navigating to question selector
Now that the user can navigate backwards, it was possible to go for Viz
(with questions already selected) back to Geography, and then next to
Questions again. This would leave the previously selected questions
checked, causing the user to have to manually adjust them. It's not
clear which use case we should support (clear them out or save them,
either is a convenience depending on the user intent). This commit clears
the selection before navigating forwards, but can be removed if the
opposite behavior is better received.