Allow Saved Visualizations
Users are able to click 'save' while on the visualizations page,
and store their currently selected geographies/questions/etc. into
local storage.

Data is stored in an array of objects, with each object containing a
title and the information needed to recreate the visualization. The
stored visualization does NOT include the survey data.