Allow downloading charts as PNG
All of the Nivo chart types implemented in the app render to Canvas. As
a result, we can generate a dataUrl for each canvas element and offer it
as a direct image download. A few workarounds are needed to accomplish
this:

- A ref to a particular container element is required in order to
generically search for a child canvas element
- Waffle charts are really a sequence of three separate charts that
should be treated as one
- Nivo charts have a transparent background, and a solid color needs to
be applied before writing

As a result of these extra requirements, there is some manual
manipulation of the canvas element to represent enough chart context to
be meaningful. Some content, like a title or geography, is not included
on the canvas based on the chart type. In these cases, I've made the
filename the place where _all_ data shows up. As an enhancement, we may
need to include additional text directly on the chart canvas while
generating the image.