Fix Tooltips on Production
`react-spring`, which handles the tooltip movement for Nivo, has a
known issue with tooltips not correctly following the cursor - see
https://github.com/pmndrs/react-spring/issues/1078. Until a fix is
added to react-spring for this, we've added a postinstall
script to prevent this error. Several other methods (updating the
craco config file, adding patch-package) were attempted but didn't
correct the problem.