Show Chart Regions in Consistent Order
Grouped charts showed the regions in reverse order to the other charts.
This could be related to the issue with canvas charts in Nivo
that we have had in other cases where the order of items is reversed
when ordering vertically. Reverses the order of the data items for
grouped charts so that when Nivo reverses them they will end in the
desired order.