Download CSVs
Users can download selections of the data for each chart or for the
visualizations page as a whole. The chart contains the columns
Geography, Question Code, Question Text, Response Category,
Response Variable, Gender, and Value, with a varying number of rows
per chart type. The filename contains the question codes for the
requested data, ie `A.1_B.1_gender_survey.csv`.