Add Survey Notification
The survey is hidden until 30 seconds after the user has first visited
the visualizations page, at which point the survey notification will
be rendered at the top of the view. The survey notification is set to a
fixed position so that it will be visible on rendering, regardless of
if the user has scrolled down the page, and will appear on any page
which the user is on.

After the initial rendering, the survey will continue to be displayed
until the user has clicked the survey link or has closed it, after which
the survey will not appear again.

The survey status is stored in a persisted reducer to be consistent
with how we have managed other local storage data.