Open Categories in Response to Search
When the user enters a search term on the questions page, categories
are automatically opened if they contain matching questions, and closed
if they have no matching questions. Note that this means that if the
user has manually opened or closed a category, then enters or changes a
search, the expansion of the accordion will be changed, as there isn't
a reasonable way to track which categories have been manually toggled
by the user.