Update Data
The question text didn't appear for certain questions, showing only the
answer. This was the result of an error in the formatting of the
sog_agg_region.xlsx file/data.

Specifically, items c2_clean and c2_cook
had the questions/'Variable Label', "C.2 In normal circumstances, do
you take care of any of the following household chores at least half
of the time?" and "C.2 In normal circumstances, do you take care of
any of the following household chores at least half of the time?"
Each question code must have a period after the code in order for the
parsing we do to work, so they needed to read, "C.2. In normal
circumstances, do you take care of any of the following household
chores at least half of the time?" and "C.2. In normal circumstances,
do you take care of any of the following household chores at least half
of the time?"

The data has now been updated to correct the formatting, allowing these
questions to be parsed correctly.