Add year selection
Once the data has been updated to include the 2021 data, users will
need to select which year to view out of all available years.

Which questions have data available will vary on a year-to-year
basis, and as such, the year must be selected prior to selecting the
questions to view. Following the pattern of the geography selector,
a page has been added where users are able to select a year from the
available years.

Based on the available years in the data, a list of years with radio
buttons are rendered, allowing users to select a year to view.

In the future, we will enable a multi-year comparison view, and will
enable users to select multiple years; when this occurs, the Radio and
RadioGroup components from ChakraUI should be switched out for the
Checkbox and CheckboxGroup components. To make this a drop-in
replacement, the `currentYear` prop has been updated to an array of
`currentYears`.