Show year in PNG and CSV downloads
Now that multiple years are available, users should be able to see in
their downloads which year is being displayed in PNGs and CSVs.

Adds the year to each response row in the CSV downloads.

Adds the year to the title of the waffle chart, which will only ever
display one year. Adds the year beside each bar in the bar charts (which
will show multiple years in comparison view).