Update layout
We removed the map on the main page to address boundary / border
concerns. In order to better maintain a appealing homepage, we then
adjusted the layout on the homepage to use multiple columns for the
countries and regions.

We also applied some minor adjustments to the layout throughout the
site for consistency and accessibility, adjusting text for clarity,
adding placeholder text to inputs, and moving the search inputs to be
inline with page text.