Fix horizontal chart scrolling
After upgrading ChakraUI, the menu for the download menu was rendered
but hidden prior to being opened. This caused a gap the width of the
menu to appear to the right of the charts, resulting in charts scrolling
horizontally even on full-screen (charts are expected to scroll on
mobile only).

Rendering the menu list in a portal prevents the menu from effecting
the page layout, fixing the horizontal scroll.