Update Question Selector Text
There was some text on the question selection page which related to a
2020-specific link which has been removed.

In the accordions on the question selection page, questions are listed
multiple times in order to allow users to select each question/answer
combination individually. Instead of referring to the number of items
available as 'number of questions', the text has been updated to
refer to these as 'number of options' to more accurately describe the
available items.