Handle invalid geo mode
The geography mode (region or country) must be selected for the app to
successfully access the data and work as expected. This value is set to
countries by default, but in the case of an invalid value somehow
entering the system, the app should handle it gracefully.

If an invalid value is present, the app will reset the value to
countries, and prevent content from loading until a valid value is
found.