View 'out of ten' charts in comparative view
Some questions contain the phrase 'out of ten of your neighbors...'
These questions are formatted as WaffleCharts when viewed as a
single-year chart. However, because each piece of data must be
rendered as a separate waffle chart, this isn't conducive for
comparative view. When viewed with multiple years selected, these
questions are instead displayed as a GroupedBarChart with the maximum
value set to 10.