Add build tools in dataproc container
When pre-built wheels are not available for certain architectures,
like the new M1 Apple Silicons, they need to be built locally in
the container. The Python Slim container does not come with build
tools by default.

This adds the build tools before installing dependencies, then
removes them so the image stays slim.